requirements.txt
faiss-cpu
numpy
transformers
torch
sentence-transformers
pypdf
flask
streamlit